add referral tree

+++ app/Http/Controllers/UserController.php
@@ -33,32 +33,13 @@ class UserController extends Controller
                 $query->select('user_id')->from('role_users');
             })
             ->get();
-        foreach ($users as $key => $user) {
-            $user->referral_tree = [];
-            if( $this->getParent($user) != '' ) {
-                $data = $this->getParent($user);
-                if($data != '') {
-                    $data1 = $this->getParent($data);
-                    if( $data1 != '') {
-
-                    } else {
-                        $result = [];
-                        array_push($result, $data1);
-                        array_push($result, $data);
-                        $user->referral_tree = $result;
-                    }
-                } else {
-                    $user->referral_tree = $this->showTree($this->getParent($user));
-                }
-            } else {
-                $user->referral_tree = $this->showTree($user);
-            }
-        }
-
         return view('manager_user', compact('users'));
     }
 
     public function delete($id = 0) {
+        if(!$this->isAdmin()) {
+            return abort(404);
+        }
         $user = User::find($id);
         $user->orders()->delete();
         $user->histories()->delete();
@@ -66,18 +47,54 @@ class UserController extends Controller
         return redirect('user')->with('success', 'User ' . $id . ' has been  deleted successfully!');
     }
 
-    public function showTree($user)
+    public function showTree($id)
     {
-        $result = [
-            "id" =>  $user->email,
-            "parent" => $user->email_referral,
-            "text" => (array)$user->email,
-        ];
-        return $result;
-    }
-
-    public function getParent($user) {
-        $data = User::where('email', '=', $user->email_referral)->first();
-        return $data;
+        if(!$this->isAdmin()) {
+            return abort(404);
+        }
+        $user = User::find($id);
+        $trees = [];
+        if( $user != '' ) {
+            $parent = $this->getParent($user);
+            if( $parent != '' ) {
+                $parent2 = $this->getParent($parent);
+                if( $parent2 != '' ) {
+                    $parent3 = $this->getParent($parent2);
+                    if( $parent3 != '' ) {
+                        array_push($trees, $parent3);
+                        array_push($trees, $parent2);
+                        array_push($trees, $parent);
+                        array_push($trees, $user);
+                    } else {
+                        array_push($trees, $parent2);
+                        array_push($trees, $parent);
+                        array_push($trees, $user);
+                    }
+                } else {
+                    array_push($trees, $parent);
+                    array_push($trees, $user);
+                }
+            } else {
+                array_push($trees, $user);
+            }
+            $children = $this->getChildren($user);
+            if( $children != '' ) {
+                $children2 = $this->getChildren($children);
+                if( $children2 != '' ) {
+                    $children3 = $this->getChildren($children2);
+                    if( $children3 != '' ) {
+                        array_push($trees, $children3);
+                        array_push($trees, $children2);
+                        array_push($trees, $children);
+                    } else {
+                        array_push($trees, $children2);
+                        array_push($trees, $children);
+                    }
+                } else {
+                    array_push($trees, $children);
+                }
+            }
+        }
+        return view('user_show_tree', compact('trees', 'user'));
     }
 }
